// app/interface/interface.ts
export interface VerticalLineProps {
  class?: string;
  height?: string;
}
export interface TitleProps {
  class?: string;
  fontSize?: string;
  title?: string;
  subTitle?: string;
  topTitle?: string;
  fontWeight?: string;
}
export interface IconItemProps {
  class?: string;
  icon: string;
  content: string | null;
  width?: string;
  iconSize?: string;
  fontSize?: string;
}

export interface TimeSectionProps {
  class?: string;
  jobTitle?: string;
  companyName?: string;
  content?: string;
  lineHeight?: string;
}